+++ src/ed.js
@@ -53,6 +53,7 @@ export default class Ed {
     this.onShareUrl = options.onShareUrl
     this.onShareFile = options.onShareFile
     this.onPlaceholderCancel = options.onPlaceholderCancel || noop
+    this.onCommandsChanged = options.onCommandsChanged
 
     // Listen for first render
     this.on('plugin.widget.initialized', options.onMount || noop)
@@ -170,6 +171,12 @@ export default class Ed {
       }
       this._content[block.id] = block
     }
+  }
+  execCommand (commandName) {
+    if (!this.pm || !this.pm.commands || !this.pm.commands[commandName] || !this.pm.commands[commandName].exec) {
+      throw new Error('Can not exec this commandName: ' + commandName)
+    }
+    this.pm.commands[commandName].exec(this.pm)
   }
   on (eventName, func) {
     let events = this._events[eventName]